Implements many core features.
This commit contains, by and large, most of the code that I wrote for the
discord hentai bot a while back. The RequestHandler logic is now decoupled
from the Hentai class. I also added a convert method for the Extension enum
and a couple of useful default params. Most importantly, this also adds an
lazy implementation for the get_homepage() method. What is does is fetching
all id's from the resulting response, and uses this information to call the
constructor. This is not very efficient since the initial response already
contains all meta data, but I don't have an alternative way to instantiate
objects offline.

Food for thoughts: Maybe add static methods as offline counterpart to the
already existing properties? This is an easy fix for an another time.