[DEV] Completely includes speedtest.py as dependency in /dev.
After countless attempts, I could not get this to work with a pip install.
So I grabbed the the source file and added it to the repository. It is
licensed under Apache 2, and according to
<http://www.apache.org/licenses/GPL-compatibility.html>
this import direction is compatible with GPLv3. For more information see
also the preamble of the source file.