[ENH] Adds random chrome user agent to the request session.
I decided that it would be best to use the faker.providers.user_agent
sub-package from the faker library to handle the random user agent string
generation for the time being. It does introduce a new external dependency,
however, this solution is very concise.

At least once a year someone should update the chrome parameter on

https://en.wikipedia.org/wiki/Google_Chrome_version_history

to extend the version and build range. According to this article,

https://www.zdnet.com/article/document-signing-service-docsketch-discloses-security-breach/

Google intends to phase out support for UA strings very soon, other vendors
may follow suite. Therefore, this is something we have to keep an eye on.