[ENH] Refactors code base by 200 LoC.
This commit is a candidate for the last major release of this library.
It breaks backwards compatibility again, but this time I can do this in
good faith because it makes working with this library overall much easier.

It uses the overloaded constructor to get rid off (almost) all static
`get_*` methods in `Hentai`. Moreover, the Utils also benefits from this
change since query-related methods now can return lists of Hentai objects.