using kwargs instead of positional
I was getting an error on hentai.download() and after a bit of debugging I found that the use of positional arguments messes things us. The Retry initializer takes these arguments in a different order that the ones given, so changing them to keyword argument fixes the error. Also it's backward compatible.

The error I was getting was
return min(timeout) < 0:
not supported '<' between list and tuple