[ENH] Refactores code.
As announced in the last version, the static Tag method have been deprecated
in favor of a single-purpose get method which takes an property as parameter:

```python
>>> from hentai import Hentai, Tag
>>> doujin = Hentai(177013)
>>> print(Tag.get(doujin.language, property='name'))
english, translated
```

Besides, the `self.download` function now takes an `folder` instead
of `dest` as parameter, allowing the user to decide beforehand the name
of the target directory (which now also defaults to `self.id` in CWD).

Additionally, the error message now return more descriptive text.

Lastly, the former `Utils.__dictionary` function is now expose as a method
in the `Hentai` class.