[ENH] [STY] Adds new feature to Tag class and improves style.
The `Tag` class now implements a `list` method supporting the following
options:

- `Option.Artist`
- `Option.Character`
- `Option.Groups`
- `Option.Parody`
- `Option.Tag`

The JSON files this method is based on was created on December 20, 2020.
Other than that, various style fixes have been made to the file. Note that
the `self.related` and `self.thread` methods have been declared as properties.